add _append_mvp_value()
This is a helper method, mainly intended for use in plugins seeking to
manipulate the hash stored in _config -- that is, the hash of plugin options
being stored in the implementing stash.

Rather than clobber any existing MVP keys, it provides a consistent method of
appending the values given to the next free slots.